risoluzione bug
