+++ ABLV_IngSW/src/main/java/db/repository/PunteggioDAO.java
@@ -1,0 +1,13 @@
+package db.repository;
+
+import java.util.List;
+
+import model.Concorrente;
+
+public class PunteggioDAO {
+	public PunteggioDAO() {}
+	
+	public void registraAssenti(String codiceGara, List<Concorrente> assenti) {
+		
+	}
+}
