+++ ABLV_IngSW/src/main/java/service/PunteggioService.java
@@ -1,0 +1,16 @@
+package service;
+
+import java.util.List;
+
+import db.repository.PunteggioDAO;
+import model.Concorrente;
+
+public class PunteggioService {
+	
+	private final PunteggioDAO punteggioDAO = new PunteggioDAO();
+	public PunteggioService() {}
+		
+	public void registraAssenti(String codiceGara, List<Concorrente> assenti){
+		punteggioDAO.registraAssenti(codiceGara, assenti);
+	}
+}
